+++ packages/stargate/src/queries/auth.ts
@@ -2,7 +2,7 @@ import { QueryClientImpl } from "cosmjs-types/cosmos/auth/v1beta1/query";
 import { Any } from "cosmjs-types/google/protobuf/any";
 
 import { QueryClient } from "./queryclient";
-import { createProtobufRpcClient, toAccAddress } from "./utils";
+import { createProtobufRpcClient } from "./utils";
 
 export interface AuthExtension {
   readonly auth: {
@@ -14,16 +14,6 @@ export interface AuthExtension {
      * `typeUrl` and decode the `value` using its own type decoder.
      */
     readonly account: (address: string) => Promise<Any | null>;
-    readonly verified: {
-      /**
-       * Returns an account if it exists and `null` otherwise.
-       *
-       * The account is a protobuf Any in order to be able to support many different
-       * account types in one API. The caller needs to switch over the expected and supported
-       * `typeUrl` and decode the `value` using its own type decoder.
-       */
-      readonly account: (address: string) => Promise<Any | null>;
-    };
   };
 }
 
@@ -38,15 +28,6 @@ export function setupAuthExtension(base: QueryClient): AuthExtension {
       account: async (address: string) => {
         const { account } = await queryService.Account({ address: address });
         return account ?? null;
-      },
-      verified: {
-        account: async (address: string) => {
-          // https://github.com/cosmos/cosmos-sdk/blob/8cab43c8120fec5200c3459cbf4a92017bb6f287/x/auth/types/keys.go#L29-L32
-          const key = Uint8Array.from([0x01, ...toAccAddress(address)]);
-          const responseData = await base.queryVerified("acc", key);
-          if (responseData.length === 0) return null;
-          return Any.decode(responseData);
-        },
       },
     },
   };