+++ packages/stargate/src/queries/bank.ts
@@ -1,11 +1,10 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-import { toAscii } from "@cosmjs/encoding";
 import { assert } from "@cosmjs/utils";
 import { QueryClientImpl } from "cosmjs-types/cosmos/bank/v1beta1/query";
 import { Coin } from "cosmjs-types/cosmos/base/v1beta1/coin";
 
 import { QueryClient } from "./queryclient";
-import { createProtobufRpcClient, toAccAddress } from "./utils";
+import { createProtobufRpcClient } from "./utils";
 
 export interface BankExtension {
   readonly bank: {
@@ -13,9 +12,6 @@ export interface BankExtension {
     readonly allBalances: (address: string) => Promise<Coin[]>;
     readonly totalSupply: () => Promise<Coin[]>;
     readonly supplyOf: (denom: string) => Promise<Coin>;
-    readonly verified: {
-      readonly balance: (address: string, denom: string) => Promise<Coin | null>;
-    };
   };
 }
 
@@ -44,19 +40,6 @@ export function setupBankExtension(base: QueryClient): BankExtension {
         const { amount } = await queryService.SupplyOf({ denom: denom });
         assert(amount);
         return amount;
-      },
-      verified: {
-        balance: async (address: string, denom: string) => {
-          // balance key is a bit tricker, using some prefix stores
-          // https://github.com/cosmwasm/cosmos-sdk/blob/80f7ff62f79777a487d0c7a53c64b0f7e43c47b9/x/bank/keeper/view.go#L74-L77
-          // ("balances", binAddress, denom)
-          // it seem like prefix stores just do a dumb concat with the keys (no tricks to avoid overlap)
-          // https://github.com/cosmos/cosmos-sdk/blob/2879c0702c87dc9dd828a8c42b9224dc054e28ad/store/prefix/store.go#L61-L64
-          // https://github.com/cosmos/cosmos-sdk/blob/2879c0702c87dc9dd828a8c42b9224dc054e28ad/store/prefix/store.go#L37-L43
-          const key = Uint8Array.from([...toAscii("balances"), ...toAccAddress(address), ...toAscii(denom)]);
-          const responseData = await base.queryVerified("bank", key);
-          return responseData.length ? Coin.decode(responseData) : null;
-        },
       },
     },
   };