+++ packages/stargate/src/stargateclient.ts
@@ -218,11 +218,6 @@ export class StargateClient {
     }
   }
 
-  public async getAccountVerified(searchAddress: string): Promise<Account | null> {
-    const account = await this.forceGetQueryClient().auth.verified.account(searchAddress);
-    return account ? accountFromAny(account) : null;
-  }
-
   public async getSequence(address: string): Promise<SequenceResponse> {
     const account = await this.getAccount(address);
     if (!account) {